+++ clink/src/Vim.h
@@ -4,6 +4,6 @@
 #include <string>
 #include <vector>
 
-int vim_open(const std::string &filename, unsigned line, unsigned col);
+int vim_open(const std::string &filename, unsigned long line, unsigned long col);
 
 std::vector<std::string> vim_highlight(const std::string &filename);
